fix ci

+++ source/GraphicsPipeline.cpp
@@ -247,7 +247,7 @@ void GraphicsPipeline::createPipeline() {
 
     VkPipelineDynamicStateCreateInfo pipelineDynamicStateCreateInfo = {
         .sType             = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO,
-        .dynamicStateCount = dynamicStateEnables.size(),
+        .dynamicStateCount = static_cast<uint32_t>(dynamicStateEnables.size()),
         .pDynamicStates    = dynamicStateEnables.data(),
     };
